+++ src/main/scala/play/api/cache/redis/package.scala
@@ -3,7 +3,7 @@ package play.api.cache
 /**
   * @author Karel Cemus
   */
-package object redis extends AnyRef with Expiration {
+package object redis extends AnyRef with util.Expiration {
 
   type SynchronousResult[ A ] = A
   type AsynchronousResult[ A ] = scala.concurrent.Future[ A ]
